refactor: edit UsersService
